% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Returns the log likelihood of the data.}
\usage{
loglikelihood(theta, tab, J, K, lambda_dd)
}
\arguments{
\item{theta}{A numeric vector of length  Jx(J^2+K) describing the joint distribution of the data.}

\item{tab}{A dataframe or a list of dataframes containing tabulated data or a list of tabulated data split by controls. The columns should be numeric with names \code{Y1}, \code{Y2}, \code{X}, and \code{n} where \code{Y1} and \code{Y2} take each value between 1 and \code{J}, \code{X} takes each value between \code{1} and \code{K}, and}

\item{J}{An integer or list corresponding to the number of unique values of \code{Y1} and \code{Y2}.}

\item{K}{An integer or list corresponding to the number of unique values of \code{X}.}

\item{lambda_dd}{A numeric value scaling a violations of diagonal dominance for Delta.}
}
\value{
A numeric value: the log likelihood of the data given \code{theta}
(i.e. given \code{Pi} and \code{Delta}), less the diagonal-dominance penalty.
}
\description{
Returns the log likelihood of the data.
}
\examples{
set.seed(1)
syn <- synthetic_data(J = 3, K = 3, I = 1, sample_size = 1000)
theta <- c(model_to_Pi_NP(rep(0, 8), J = 3),
           model_to_Delta_NP_ind(rep(0, 12)))
loglikelihood(theta, syn$tab[[1]], J = 3, K = 3, lambda_dd = 0)
}
\keyword{internal}
