% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_Pi_NP.R
\name{model_to_Pi_NP}
\alias{model_to_Pi_NP}
\title{Maps model parameters, phi, to the joint distribution of X and Y*, Pi.}
\usage{
model_to_Pi_NP(phi, J, ...)
}
\arguments{
\item{phi}{A numeric vector.}

\item{J}{An integer corresponding to the dimension of Y.}

\item{...}{Additional, optional arguments.}
}
\value{
A numeric vector of length \code{J * K} corresponding to the \code{J} by \code{K}
matrix \code{Pi}, stored column-major: rows index \eqn{Y^*}, columns index \eqn{X}.
}
\description{
Maps model parameters, phi, to the joint distribution of X and Y*, Pi.
}
\examples{
# phi = log(2, 3, 4) maps to probabilities (2, 3, 4, 1) / 10
model_to_Pi_NP(log(c(2, 3, 4)), J = 2)

# Flat parameters give the uniform joint distribution
matrix(model_to_Pi_NP(rep(0, 3 * 3 - 1), J = 3), nrow = 3)
}
