% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pi_to_betas.R
\name{Pi_to_betas}
\alias{Pi_to_betas}
\title{Maps the joint distribution, Pi, of X and Y* to a vector, representing beta in each covariate cell}
\usage{
Pi_to_betas(Pi, X_vals, Y_vals)
}
\arguments{
\item{Pi}{A list of numeric vectors containing the elements of Pi.}

\item{X_vals}{A list of numeric vectors representing the scalar values associated with X.}

\item{Y_vals}{A list of numeric vectors representing the scalar values associated with Y.}
}
\value{
A scalar equal to beta.
}
\description{
Longer description of what it does...
}
\details{
some details.
}
\examples{
\dontrun{
some example code # Should return something
}
}
