% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_prior_Delta_NP_ind.R
\name{log_prior_Delta_NP_ind}
\alias{log_prior_Delta_NP_ind}
\title{Evaluates the log of the prior of Delta at model_to_Delta_NP_ind(psi).}
\usage{
log_prior_Delta_NP_ind(psi)
}
\arguments{
\item{psi}{A numeric vector parameterizing \code{Delta} through \code{model_to_Delta_NP_ind}.}
}
\value{
A numeric value equal to the log of the flat prior of \code{Delta} at \code{psi}, re-scaled for the logit transform.
}
\description{
Evaluates the log of the prior of Delta at model_to_Delta_NP_ind(psi).
}
