% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_link_volume.R
\name{logit_link_volume}
\alias{logit_link_volume}
\title{Evaluates the log of the determinant of the Jacobian of the logit link.}
\usage{
logit_link_volume(x)
}
\arguments{
\item{x}{A numeric vector in logit-space corresponding to some .}
}
\value{
A numeric value equal to the log of the determinant of the Jacobian of the logit transform.
}
\description{
Evaluates the log of the determinant of the Jacobian of the logit link.
}
\keyword{internal}
