% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Returns the log likelihood of the data.}
\usage{
loglikelihood(theta)
}
\arguments{
\item{theta}{A numeric vector of length  Jx(J^2+K) describing the joint distribution of the data.}
}
\value{
the log likelihood of the data given theta, i.e. Pi and Delta.
}
\description{
Returns the log likelihood of the data.
}
