% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_empirical_Delta_RL.R
\name{make_empirical_Delta_RL}
\alias{make_empirical_Delta_RL}
\title{Creates a function for model_to_Delta based on RL structure errors and the empirical distribution of Y1 and Y2}
\usage{
make_empirical_Delta_RL(tab, J)
}
\arguments{
\item{tab}{tab A dataframe or a list of dataframes containing tabulated data or a list of tabulated data split by controls. The columns should have names \code{Y1}, \code{Y2}, \code{X}, and \code{n} where \code{n} is a non-negative numeric vector corresponding to the counts of \code{Y1},\code{Y2}, and \code{X}. The rows should be ordered according to \code{order(Y2,Y1,X)}.}

\item{J}{An integer or list corresponding to the number of unique values of \code{Y1} and \code{Y2}.}
}
\value{
A list including 1. a function or a list of functions that map model parameters \code{psi} to the misclassification matrix \code{Delta}, 2. a vector or list of vectors corresponding to initial values of psi , and 3. a function or list of functions for the log prior of Delta in this model.
}
\description{
Creates a function for model_to_Delta based on RL structure errors and the empirical distribution of Y1 and Y2
}
