% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misclassifyr.R
\name{misclassifyr}
\alias{misclassifyr}
\title{misclassifyr}
\usage{
misclassifyr(
  tab,
  J,
  K,
  model_to_Pi = model_to_Pi_NP,
  model_to_Delta = model_to_Delta_NP_ind,
  phi_0 = NA,
  psi_0 = NA,
  X_names = NA,
  Y_names = NA,
  W_names = NA,
  estimate_beta = F,
  estimate_betas = F,
  X_vals = NA,
  Y_vals = NA,
  optim_tol = 1e-08,
  optim_maxit = 1e+05,
  bayesian = F,
  prior_Pi = prior_Pi_NP,
  prior_Delta = prior_Delta_NP_ind,
  n_mcmc_draws = 20000,
  n_burnin = 10000,
  thinning_rate = 2,
  gibbs_proposal_sd = 0.1,
  cores = 1
)
}
\arguments{
\item{tab}{A dataframe or a list of dataframes containing tabulated data or a list of tabulated data split by controls. The columns should be numeric with names \code{Y1}, \code{Y2}, \code{X}, and \code{n} where \code{Y1} and \code{Y2} take each value between \code{1} and \code{J}, \code{X} takes each value between \code{1} and \code{K}, and}

\item{J}{An integer or list corresponding to the number of unique values of \code{Y1} and \code{Y2}.}

\item{K}{An integer or list corresponding to the number of unique values of \code{X}.}

\item{model_to_Pi}{A function or list of functions mapping the parameters of a model for the joint distribution to the joint distribution \verb{\eqn{\Pi}}.}

\item{model_to_Delta}{A function or list of functions mapping the parameters of a model to the conditional distribution Y1, Y2 | Y*, \verb{\eqn{\Delta}}.}

\item{phi_0}{A numeric vector or list of numeric vectors providing the starting location for optimization for the argument to model_to_Pi.}

\item{psi_0}{A numeric vector or list of numeric vectors providing the starting location for optimization for the argument to model_to_Delta.}

\item{X_names}{A character vector or list corresponding to the values of the regressor X.}

\item{Y_names}{A character vector or list corresponding to the values of the outcome Y.}

\item{W_names}{A character vector corresponding to the values of the control W in each cell.}

\item{estimate_beta}{A logical value indicating whether to regress Y on X.}

\item{estimate_betas}{A logical value indicating whether to regress Y on X within covariate cells.}

\item{X_vals}{A numeric vector or list of numeric vectors providing the values of X associated with the columns of Pi.}

\item{Y_vals}{A numeric vector or list of numeric vectors providing the values of Y associated with the rows of Pi.}

\item{optim_tol}{A numeric value giving the relative tolerance for optimization with the optim.}

\item{optim_maxit}{An integer giving the maximum number of iterations for optim.}

\item{bayesian}{A logical value indicating whether or not to compute the posterior of values.}

\item{n_mcmc_draws}{An integer corresponding to the length of the MCMC chain.}

\item{n_burnin}{An integer giving the length of the burn-in period for each MCMC chain, must be shorter than \code{n_mcmc_draws}.}

\item{thinning_rate}{An integer indicating how frequently to record posterior draws from the MCMC chain -- e.g. a \code{thinning_rate} of 2 records every other draw.}

\item{gibbs_proposal_sd}{A numeric value giving the standard deviation for the proposal distribution in each Gibbs step.}

\item{cores}{An integer for the number of CPUs available for parallel processing.}

\item{split_eta}{An integer or list indicating where to split the vector \code{eta} in \code{phi} and \code{psi}, the arguments to \code{model_to_Pi} and \code{model_to_Delta} respectively.}

\item{log_prior_Pi}{A function or list of functions evaluating the log of the prior of Pi at \code{phi} (in logs!).}

\item{log_prior_Delta}{A function or list of functions evaluating the log of the prior of Delta at \code{psi} (in logs!).}
}
\value{
An object that includes estimates and information from the estimation process
}
\description{
This function provides a menu of options for estimation and inference of misclassification models in which the analyst has access to two noisy measures, \code{Y1} and \code{Y2} of a latent outcome \verb{Y*}, a correctly measured covariate \code{X}, and discrete controls \code{W}.
}
