% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_Delta_NP.R
\name{model_to_Delta_NP}
\alias{model_to_Delta_NP}
\title{Maps model parameters, psi, to the conditional distribution Y1, Y2 | Y*, Delta.}
\usage{
model_to_Delta_NP(psi)
}
\arguments{
\item{psi}{A numeric vector containing Pi, Delta^{(1)}, and Delta^{(2)}.}
}
\value{
A numeric vector corresponding to the (JxJ)xJ matrix Delta.
}
\description{
Maps model parameters, psi, to the conditional distribution Y1, Y2 | Y*, Delta.
}
