% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_Delta_NP_ind.R
\name{model_to_Delta_NP_ind}
\alias{model_to_Delta_NP_ind}
\title{Maps model parameters, psi, to the conditional distribution Y1, Y2 | Y*, Delta.}
\usage{
model_to_Delta_NP_ind(psi)
}
\arguments{
\item{psi}{A numeric vector of length 2xJx(J-1) containing Delta^{(1)} and Delta^{(2)}.}
}
\value{
A numeric vector corresponding to the (JxJ)xJ matrix Delta. d
}
\description{
Maps model parameters, psi, to the conditional distribution Y1, Y2 | Y*, Delta.
}
