% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_Delta_RL_ind.R
\name{model_to_Delta_RL_ind}
\alias{model_to_Delta_RL_ind}
\title{Maps model parameters, psi, to the joint distribution of the data, theta.}
\usage{
model_to_Delta_RL_ind(psi)
}
\arguments{
\item{psi}{A numeric vector of length 2(J-1)+2J corresponding to the column and row scales of the record linkage.}
}
\value{
something
}
\description{
Longer description of what it does...
}
\details{
some details.
}
\examples{
\dontrun{
some example code # Should return something
}
}
