% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_Pi_NP.R
\name{model_to_Pi_NP}
\alias{model_to_Pi_NP}
\title{Maps model parameters, phi, to the joint distribution of X and Y*, Pi.}
\usage{
model_to_Pi_NP(phi)
}
\arguments{
\item{phi}{A numeric vector.}

\item{J}{An integer corresponding to the dimension of Y.}

\item{K}{An integer corresponding to the dimension of X.}
}
\value{
something
}
\description{
Longer description of what it does...
}
\details{
some details.
}
\examples{
\dontrun{
some example code # Should return something
}
}
