% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pi_to_beta.R
\name{Pi_to_beta}
\alias{Pi_to_beta}
\title{Estimate a linear regression coefficient \eqn{\beta} from the joint distribution Pi}
\usage{
Pi_to_beta(
  X_vals,
  Y_vals,
  W_weights = NA,
  W_names = NA,
  mle = T,
  bayesian = F,
  Pi_mle = NA,
  cov_Pi = NA,
  posterior_Pi = NA,
  ll_history = NA
)
}
\arguments{
\item{X_vals}{A numeric vector or list of numeric vectors providing the values of X associated with the columns of Pi.}

\item{Y_vals}{A numeric vector or list of numeric vectors providing the values of Y associated with the rows of Pi.}

\item{W_weights}{A numeric vector representing the sample size of each control cell.}

\item{W_names}{A character vector corresponding to the values of the control \eqn{W} in each cell.}

\item{mle}{Logical value indicating whether maximum likelihood estimates of \eqn{\Pi} have been provided. Defaults to \code{TRUE}.}

\item{bayesian}{Logical value indicating whether posterior draws of \eqn{\Pi} have been provided. Defaults to \code{FALSE}.}

\item{Pi_mle}{A numeric vector or list of numeric vectors containing the elements of Pi.}

\item{cov_Pi}{A numeric vector or a list of numeric vectors representing the covariance of estimates of the elements of Pi.}

\item{posterior_Pi}{A data frame of posterior draws of Pi (or a list of such data frames split by control cell), as returned in \verb{$posterior_Pi} by \code{misclassifyr(bayesian = TRUE)}.}

\item{ll_history}{A data frame with columns \code{ll} and \code{draw} (or a list of such data frames split by control cell), as returned in \verb{$ll_history} by \code{misclassifyr(bayesian = TRUE)}. Required for the Chen-Christensen-Tamer HPD confidence set; if omitted, \code{HPD_draws} and \code{HPDCI} are returned as \code{NA}.}
}
\description{
This function computes \eqn{\beta} in the linear regression \eqn{Y^* = \alpha + \beta X + \epsilon} given the joint distribution of \eqn{X} and \eqn{Y^*}, \eqn{\Pi}, and a set of scalars associated with each value of \eqn{X} and \eqn{Y^*}.
}
