% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pi_to_beta.R
\name{Pi_to_beta}
\alias{Pi_to_beta}
\title{Estimate a linear regression coefficient \eqn{\beta} from the joint distribution Pi}
\usage{
Pi_to_beta(
  X_vals,
  Y_vals,
  W_weights = NA,
  W_names = NA,
  mle = T,
  bayesian = F,
  Pi_mle = NA,
  cov_Pi = NA,
  posterior_Pi = NA
)
}
\arguments{
\item{X_vals}{A numeric vector or list of numeric vectors providing the values of X associated with the columns of Pi.}

\item{Y_vals}{A numeric vector or list of numeric vectors providing the values of Y associated with the rows of Pi.}

\item{W_weights}{A numeric vector representing the sample size of each control cell.}

\item{W_names}{A character vector corresponding to the values of the control \eqn{W} in each cell.}

\item{mle}{Logical value indicating whether maximum likelihood estimates of \eqn{\Pi} have been provided. Defaults to \code{TRUE}.}

\item{bayesian}{Logical value indicating whether posterior draws of \eqn{\Pi} have been provided. Defaults to \code{FALSE}.}

\item{Pi_mle}{A numeric vector or list of numeric vectors containing the elements of Pi.}

\item{cov_Pi}{A numeric vector or a list of numeric vectors representing the covariance of estimates of the elements of Pi.}
}
\description{
This function computes \eqn{\beta} in the linear regression \eqn{Y^* = \alpha + \beta X + \epsilon} given the joint distribution of \eqn{X} and \eqn{Y^*}, \eqn{\Pi}, and a set of scalars associated with each value of \eqn{X} and \eqn{Y^*}.
}
