% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_misclassification_data.R
\name{prep_misclassification_data}
\alias{prep_misclassification_data}
\title{prep_misclassification_data}
\usage{
prep_misclassification_data(
  data,
  outcome_1,
  outcome_2,
  regressor,
  controls = NA,
  weights = NA,
  record_vals = F
)
}
\arguments{
\item{data}{A data.frame containing the outcome variable,}

\item{outcome_1}{A character string denoting the variable in the dataframe to be used as the first measure of an outcome, Y_1.}

\item{outcome_2}{A character string denoting the variable in the dataframe to be used as the second measure of an outcome, Y_2.}

\item{regressor}{A character string denoting the variable in the dataframe to be used as a regressor, X.}

\item{controls}{A character string or vector of character strings denoting the variable/variables to be used as non-parametric controls, W.}

\item{weights}{A character string denoting a variable containing individual level weights}

\item{record_vals}{A logical value indicating whether to record the unique values of the outcomes and the regressor. If record_vals = F, you likely want to order the data by the regressor and outcomes before applying prep_misclassification_data.}
}
\value{
A list of objects including tabulated data to be used in misclassifyr()
}
\description{
This function tabulates data and generates metadata in a format to be used with the misclassifyr() function.
}
