% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_beta_deltamethod.R
\name{se_beta_deltamethod}
\alias{se_beta_deltamethod}
\title{Computes the standard error of beta as function of the joint distribution of X and Y*, Pi}
\usage{
se_beta_deltamethod(Pi, cov_Pi, X_vals, Y_vals, W_weights)
}
\arguments{
\item{Pi}{A numeric vector or list of numeric vectors containing the elements of Pi.}

\item{cov_Pi}{A numeric vector or a list of numeric vectors representing the covariance of estimates of the elements of Pi.}

\item{X_vals}{A numeric vector or a list of numeric vectors representing the scalar values associated with X.}

\item{Y_vals}{A numeric vector or a list of numeric vectors representing the scalar values associated with Y.}

\item{W_weights}{A numeric vector representing the sample size of each control cell.}
}
\value{
A scalar equal to the standard error of beta.
}
\description{
Computes the standard error of beta as function of the joint distribution of X and Y*, Pi
}
