% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softlog.R
\name{softlog}
\alias{softlog}
\title{Logarithm with a lower bound}
\usage{
softlog(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector composed of the elements of \code{log(x)} or \code{log(1e-20)} for element is less than \code{1e-20}.
}
\description{
Logarithm with a lower bound
}
\keyword{internal}
