% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_data.R
\name{synthetic_data}
\alias{synthetic_data}
\title{Generates synthetic misclassification data.}
\usage{
synthetic_data(
  J = 5,
  K = 5,
  I = 2,
  sample_size = 1e+06,
  dgp_delta = "Nonparametric, independent, strong diagonal",
  dgp_pi = "Exponential"
)
}
\arguments{
\item{J}{An integer indicating the dimension of Y.}

\item{K}{An integer indicating the dimension of X.}

\item{I}{An integer indicating the dimension of W.}

\item{sample_size}{An integer denoting the number of synthetic observations.}

\item{dgp_delta}{A character string indicating the data generating process for the synthetic noise}

\item{dgp_pi}{A character string indicating the data generating process for the joint distribution of X and Y*}
}
\value{
A list including tabulated data \code{tab} and matrices \code{Pi}, \code{Delta}
}
\description{
Longer description of what it does...
}
\details{
some details...
}
\examples{
\dontrun{
some example code # Should return something
}
}
